--- a/.travis.yml
+++ b/.travis.yml
@@ -15,12 +15,11 @@ jobs:
         packages:
         - elfutils
         - libzstd1-dev
-        - libb2-dev
 
   # Job 2: Build on Linux with native 32-bit GCC
   - os: linux
     compiler: gcc
-    env: T="Linux GCC 32-bit" VERBOSE=1 CFLAGS="-m32 -g -O2" CXXFLAGS="-m32 -g -O2" LDFLAGS="-m32" CMAKE_PARAMS="-DUSE_LIBZSTD_FROM_INTERNET=ON -DUSE_LIBB2_FROM_INTERNET=ON" ENABLE_CACHE_CLEANUP_TESTS=1
+    env: T="Linux GCC 32-bit" VERBOSE=1 CFLAGS="-m32 -g -O2" CXXFLAGS="-m32 -g -O2" LDFLAGS="-m32" CMAKE_PARAMS="-DUSE_LIBZSTD_FROM_INTERNET=ON" ENABLE_CACHE_CLEANUP_TESTS=1
     addons:
       apt:
         packages:
@@ -37,7 +36,6 @@ jobs:
         packages:
         - elfutils
         - libzstd1-dev
-        - libb2-dev
 
   # Job 4: Build on Linux with native GCC and CUDA compiler
   - os: linux
@@ -48,14 +46,13 @@ jobs:
         packages:
         - elfutils
         - libzstd1-dev
-        - libb2-dev
     before_install:
     - source ./misc/install_cuda.sh
 
   # Job 5: Build on Linux with 32-bit MinGW cross-compiler
   - os: linux
     compiler: i686-w64-mingw32-g++-posix
-    env: T="Linux MinGW 32-bit" VERBOSE=1 CC=i686-w64-mingw32-gcc-posix CXX=i686-w64-mingw32-g++-posix CMAKE_PARAMS="-DCMAKE_SYSTEM_NAME=Windows -DUSE_LIBZSTD_FROM_INTERNET=ON -DUSE_LIBB2_FROM_INTERNET=ON" RUN_TESTS=0 ENABLE_CACHE_CLEANUP_TESTS=1
+    env: T="Linux MinGW 32-bit" VERBOSE=1 CC=i686-w64-mingw32-gcc-posix CXX=i686-w64-mingw32-g++-posix CMAKE_PARAMS="-DCMAKE_SYSTEM_NAME=Windows -DUSE_LIBZSTD_FROM_INTERNET=ON" RUN_TESTS=0 ENABLE_CACHE_CLEANUP_TESTS=1
     addons:
       apt:
         packages:
@@ -64,7 +61,7 @@ jobs:
   # Job 6: Build on Linux with 64-bit MinGW cross-compiler
   - os: linux
     compiler: x86_64-w64-mingw32-g++
-    env: T="Linux MinGW 64-bit" VERBOSE=1 CC=x86_64-w64-mingw32-gcc-posix CXX=x86_64-w64-mingw32-g++-posix CMAKE_PARAMS="-DCMAKE_SYSTEM_NAME=Windows -DUSE_LIBZSTD_FROM_INTERNET=ON -DUSE_LIBB2_FROM_INTERNET=ON" RUN_TESTS=0 ENABLE_CACHE_CLEANUP_TESTS=1
+    env: T="Linux MinGW 64-bit" VERBOSE=1 CC=x86_64-w64-mingw32-gcc-posix CXX=x86_64-w64-mingw32-g++-posix CMAKE_PARAMS="-DCMAKE_SYSTEM_NAME=Windows -DUSE_LIBZSTD_FROM_INTERNET=ON" RUN_TESTS=0 ENABLE_CACHE_CLEANUP_TESTS=1
     addons:
       apt:
         packages:
@@ -73,7 +70,7 @@ jobs:
   # Job 7: Build on MacOS with native Clang
   - os: osx
     compiler: clang
-    env: T="MacOS Clang" VERBOSE=1 CMAKE_PARAMS="-DUSE_LIBZSTD_FROM_INTERNET=ON -DUSE_LIBB2_FROM_INTERNET=ON" ENABLE_CACHE_CLEANUP_TESTS=1
+    env: T="MacOS Clang" VERBOSE=1 CMAKE_PARAMS="-DUSE_LIBZSTD_FROM_INTERNET=ON" ENABLE_CACHE_CLEANUP_TESTS=1
 
   # Job 8: Run Clang's undefined behavior sanitizer
   - os: linux
@@ -84,7 +81,6 @@ jobs:
         packages:
         - elfutils
         - libzstd1-dev
-        - libb2-dev
 
   # Job 9: Run Clang's address sanitizer
   - os: linux
@@ -95,7 +91,6 @@ jobs:
         packages:
         - elfutils
         - libzstd1-dev
-        - libb2-dev
 
   # Job 10: Run Clang's static analyzer
   - os: linux
@@ -105,7 +100,6 @@ jobs:
       apt:
         packages:
         - libzstd1-dev
-        - libb2-dev
 
   # Job 11: Run Clang-Tidy
   - os: linux
@@ -115,7 +109,6 @@ jobs:
       apt:
         packages:
         - libzstd1-dev
-        - libb2-dev
 
   # Job 12: Run Clang-Format in check mode
   - os: linux
@@ -125,7 +118,6 @@ jobs:
       apt:
         packages:
         - libzstd1-dev
-        - libb2-dev
     before_install:
     - pip install --user cmake_format
 
@@ -137,7 +129,6 @@ jobs:
         packages:
         - elfutils
         - libzstd1-dev
-        - libb2-dev
 
   # Job 14: Build source package, extract it out of source, run tests.
   - os: linux
@@ -147,7 +138,6 @@ jobs:
         packages:
         - elfutils
         - libzstd1-dev
-        - libb2-dev
 
   # Job 15: Build documentation
   - os: linux
@@ -157,7 +147,6 @@ jobs:
         packages:
         - elfutils
         - libzstd1-dev
-        - libb2-dev
         - asciidoc
 
   # Job 16: Build manpage
@@ -168,7 +157,6 @@ jobs:
         packages:
         - elfutils
         - libzstd1-dev
-        - libb2-dev
         - asciidoc
         - xsltproc
 
@@ -183,7 +171,6 @@ jobs:
       apt:
         packages:
         - libzstd1-dev
-        - libb2-dev
 
   # Job 12: Run Clang-Format check mode
   - os: linux
@@ -193,6 +180,5 @@ jobs:
       apt:
         packages:
         - libzstd1-dev
-        - libb2-dev
     before_install:
     - pip install --user cmake_format