--- a/CMakeLists.txt
+++ b/CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required(VERSION 3.4.3)
 
-project(ccache LANGUAGES C CXX)
+project(ccache LANGUAGES ASM C CXX)
 set(CMAKE_PROJECT_DESCRIPTION "a fast C/C++ compiler cache")
 
 set(CMAKE_CXX_STANDARD 11)