--- a/dockerfiles/debian-10-buster/Dockerfile
+++ b/dockerfiles/debian-10-buster/Dockerfile
@@ -14,7 +14,6 @@ RUN apt-get update && apt-get install -y --no-install-recommends \
                 cmake \
                 elfutils \
                 gcc-multilib \
-                libb2-dev \
                 libzstd-dev \
                 xsltproc \
         && rm -rf /var/lib/apt/lists/*