--- a/test-all-systems.sh
+++ b/test-all-systems.sh
@@ -28,11 +28,11 @@ echo "Warning: Docker support is rather experimental\n"
 
 # Ubuntu (ancient, old and latest)
 
-  # zstd and libb2 not available for Ubuntu 14.
-  CC=gcc   CXX=g++     CMAKE_PARAMS="-DUSE_LIBZSTD_FROM_INTERNET=ON -DUSE_LIBB2_FROM_INTERNET=ON" ./docker.sh ubuntu-14-trusty
+  # zstd not available for Ubuntu 14.
+  CC=gcc   CXX=g++     CMAKE_PARAMS="-DUSE_LIBZSTD_FROM_INTERNET=ON" ./docker.sh ubuntu-14-trusty
 
   # See https://github.com/ccache/ccache/issues/601
-  #CC=clang CXX=clang++ CMAKE_PARAMS="-DUSE_LIBZSTD_FROM_INTERNET=ON -DUSE_LIBB2_FROM_INTERNET=ON" ./docker.sh ubuntu-14-tusty
+  #CC=clang CXX=clang++ CMAKE_PARAMS="-DUSE_LIBZSTD_FROM_INTERNET=ON" ./docker.sh ubuntu-14-tusty
 
   CC=gcc   CXX=g++     ./docker.sh ubuntu-16-xenial
   CC=clang CXX=clang++ ./docker.sh ubuntu-16-xenial
@@ -43,9 +43,9 @@ echo "Warning: Docker support is rather experimental\n"
 
 # Alpine (old and latest)
 
-  CC=gcc   CXX=g++     CMAKE_PARAMS="-DUSE_LIBZSTD_FROM_INTERNET=ON -DUSE_LIBB2_FROM_INTERNET=ON" ./docker.sh alpine-3.4
+  CC=gcc   CXX=g++     CMAKE_PARAMS="-DUSE_LIBZSTD_FROM_INTERNET=ON" ./docker.sh alpine-3.4
   # Clang is not capable to compile libzstd from internet before alpine 3.12 (Some SSE2 error regarding missing file emmintrin.h)
-  #CC=clang CXX=clang++ CMAKE_PARAMS="-DUSE_LIBZSTD_FROM_INTERNET=ON -DUSE_LIBB2_FROM_INTERNET=ON" ./docker.sh alpine-3.4
-  CC=clang CXX=clang++ CMAKE_PARAMS="-DUSE_LIBZSTD_FROM_INTERNET=ON -DUSE_LIBB2_FROM_INTERNET=ON" ./docker.sh alpine-3.12
-  CC=gcc   CXX=g++     CMAKE_PARAMS="-DUSE_LIBB2_FROM_INTERNET=ON" ./docker.sh alpine-3.12
-  CC=clang CXX=clang++ CMAKE_PARAMS="-DUSE_LIBB2_FROM_INTERNET=ON" ./docker.sh alpine-3.12
+  #CC=clang CXX=clang++ CMAKE_PARAMS="-DUSE_LIBZSTD_FROM_INTERNET=ON" ./docker.sh alpine-3.4
+  CC=clang CXX=clang++ CMAKE_PARAMS="-DUSE_LIBZSTD_FROM_INTERNET=ON" ./docker.sh alpine-3.12
+  CC=gcc   CXX=g++     ./docker.sh alpine-3.12
+  CC=clang CXX=clang++ ./docker.sh alpine-3.12