--- a/unittest/test_hash.cpp
+++ b/unittest/test_hash.cpp
@@ -26,7 +26,7 @@ TEST_CASE("test_known_strings")
     struct hash* h = hash_init();
     hash_string(h, "");
     CHECK(hash_result(h).to_string()
-          == "af1349b9f5f9a1a6a0404dea36dcc9499bcb25c9adc112b7cc9a93ca");
+          == "af1349b9f5f9a1a6a0404dea36dcc9499bcb25c9");
     hash_free(h);
   }
 
@@ -34,7 +34,7 @@ TEST_CASE("test_known_strings")
     struct hash* h = hash_init();
     hash_string(h, "a");
     CHECK(hash_result(h).to_string()
-          == "17762fddd969a453925d65717ac3eea21320b66b54342fde15128d6c");
+          == "17762fddd969a453925d65717ac3eea21320b66b");
     hash_free(h);
   }
 
@@ -42,7 +42,7 @@ TEST_CASE("test_known_strings")
     struct hash* h = hash_init();
     hash_string(h, "message digest");
     CHECK(hash_result(h).to_string()
-          == "7bc2a2eeb95ddbf9b7ecf6adcb76b453091c58dc43955e1d9482b194");
+          == "7bc2a2eeb95ddbf9b7ecf6adcb76b453091c58dc");
     hash_free(h);
   }
 
@@ -53,7 +53,7 @@ TEST_CASE("test_known_strings")
       "1234567890123456789012345678901234567890123456789012345678901234567890"
       "1234567890");
     CHECK(hash_result(h).to_string()
-          == "f263acf51621980b9c8de5da4a17d314984e05abe4a21cc83a07fe3e");
+          == "f263acf51621980b9c8de5da4a17d314984e05ab");
     hash_free(h);
   }
 }
@@ -65,7 +65,7 @@ TEST_CASE("hash_result_should_not_alter_state")
   hash_result(h);
   hash_string(h, " digest");
   CHECK(hash_result(h).to_string()
-        == "7bc2a2eeb95ddbf9b7ecf6adcb76b453091c58dc43955e1d9482b194");
+        == "7bc2a2eeb95ddbf9b7ecf6adcb76b453091c58dc");
   hash_free(h);
 }
 
@@ -75,9 +75,9 @@ TEST_CASE("hash_result_should_be_idempotent")
   hash_string(h, "");
   hash_result(h);
   CHECK(hash_result(h).to_string()
-        == "af1349b9f5f9a1a6a0404dea36dcc9499bcb25c9adc112b7cc9a93ca");
+        == "af1349b9f5f9a1a6a0404dea36dcc9499bcb25c9");
   CHECK(hash_result(h).to_string()
-        == "af1349b9f5f9a1a6a0404dea36dcc9499bcb25c9adc112b7cc9a93ca");
+        == "af1349b9f5f9a1a6a0404dea36dcc9499bcb25c9");
   hash_free(h);
 }
 
@@ -89,7 +89,6 @@ TEST_CASE("hash_result digest bytes")
   uint8_t expected[Digest::size()] = {
     0x7b, 0xc2, 0xa2, 0xee, 0xb9, 0x5d, 0xdb, 0xf9, 0xb7, 0xec,
     0xf6, 0xad, 0xcb, 0x76, 0xb4, 0x53, 0x09, 0x1c, 0x58, 0xdc,
-    0x43, 0x95, 0x5e, 0x1d, 0x94, 0x82, 0xb1, 0x94,
   };
   CHECK(memcmp(d.bytes(), expected, sizeof(Digest::size())) == 0);
   hash_free(h);