<!--
  Thanks for contributing to ccache! Please read
  https://github.com/ccache/ccache/blob/master/CONTRIBUTING.md#contributing-code
  before submitting the pull request.
-->
### Description ###
Replace the BLAKE2b hash function with BLAKE3 (see #503). The BLAKE3 source files are added as third_party sources and built when building ccache.

I've compared the performance on my machine by running `make CXX=/usr/bin/g++ performance`.

On master (42939f09):
```
Without ccache:                             2.6040 s (100.0000 %) (  1.0000 x)
With ccache, preprocessor mode, cache miss: 2.7191 s (104.4190 %) (  0.9577 x)
With ccache, preprocessor mode, cache hit:  0.0892 s (  3.4245 %) ( 29.2010 x)
With ccache, direct mode, cache miss:       2.7271 s (104.7250 %) (  0.9549 x)
With ccache, direct mode, cache hit:        0.0156 s (  0.5986 %) (167.0574 x)
With ccache, depend mode, cache miss:       2.6532 s (101.8877 %) (  0.9815 x)
With ccache, depend mode, cache hit:        0.0157 s (  0.6035 %) (165.6877 x)
```

This PR:
```
Without ccache:                             2.5987 s (100.0000 %) (  1.0000 x)
With ccache, preprocessor mode, cache miss: 2.6948 s (103.7001 %) (  0.9643 x)
With ccache, preprocessor mode, cache hit:  0.0879 s (  3.3843 %) ( 29.5482 x)
With ccache, direct mode, cache miss:       2.6909 s (103.5476 %) (  0.9657 x)
With ccache, direct mode, cache hit:        0.0129 s (  0.4956 %) (201.7592 x)
With ccache, depend mode, cache miss:       2.6251 s (101.0160 %) (  0.9899 x)
With ccache, depend mode, cache hit:        0.0131 s (  0.5035 %) (198.6122 x)
```

This PR but all BLAKE3 sources built with `-O3` instead of `-O2`:
```
Without ccache:                             2.6031 s (100.0000 %) (  1.0000 x)
With ccache, preprocessor mode, cache miss: 2.7248 s (104.6749 %) (  0.9553 x)
With ccache, preprocessor mode, cache hit:  0.0886 s (  3.4035 %) ( 29.3813 x)
With ccache, direct mode, cache miss:       2.7173 s (104.3844 %) (  0.9580 x)
With ccache, direct mode, cache hit:        0.0130 s (  0.4987 %) (200.5049 x)
With ccache, depend mode, cache miss:       2.6381 s (101.3422 %) (  0.9868 x)
With ccache, depend mode, cache hit:        0.0130 s (  0.4975 %) (200.9903 x)
```