@erijo 2020-01-23
There seems to be some problem building the AVX512 version on Travis (too old GCC perhaps?). I'll let it be for now, but I can have a look at it later if you're interested in merging this PR.