@erijo 2020-01-23
I did another test where I built KDAB/hotspot@a7c2ecb1 with `make -j8`. I measured the time it took to build using three different ccache versions. For each ccache version I ran the build three times: first one with an empty ccache dir and then two times with populated ccache dir.

ccache 3.7.7 from Debian
----------------------------------
```
235,79s user 19,71s system 386% cpu 1:06,03 total (empty cache)
15,16s user 3,60s system 329% cpu 5,688 total
15,26s user 3,68s system 330% cpu 5,730 total
```

ccache master (42939f09)
-----------------------------------
```
228,53s user 19,61s system 385% cpu 1:04,44 total (empty cache)
15,79s user 3,82s system 335% cpu 5,846 total
15,51s user 3,78s system 333% cpu 5,782 total
```

This PR
-----------
```
235,94s user 19,62s system 385% cpu 1:06,30 total (empty cache)
14,93s user 3,73s system 335% cpu 5,567 total
15,11s user 3,54s system 336% cpu 5,543 total
```

Based on this limited testing it seems like this version is faster than 3.7.7 which is faster than current master.