@erijo 2020-02-10
I took a look at the CI failure and the problem seems to be that GCC 5.4 generates `vmovdqu %ymm17, (%r11)` which, as far as I can tell, is an invalid instruction. This instruction is VEX encoded and can only access registers ymm0 - ymm15. To access ymm16-ymm31 the instruction should be EVEX encoded.

Adding `-mavx512bw` to the build flags makes GCC generate `vmovdqu8 %ymm17, (%r11)` instead, which compiles (the 8 suffix makes it an EVEX instruction), but I can't really judge the consequences of that change.