@erijo 2020-02-13
I've now rebased this PR on the latest master and update BLAKE3 to the first release of the C implementation. Following the discussion in #503 I've also reverted my digest size change and the size is now back to 160 bits as it was before (in a separate commit to make it easy to go one way or the other).

With the latest BLAKE3 release all build errors are fixed and performance is a tiny bit better than before. On my computer I now get:
```
Without ccache:                             2.6827 s (100.0000 %) (  1.0000 x)
With ccache, preprocessor mode, cache miss: 2.7722 s (103.3368 %) (  0.9677 x)
With ccache, preprocessor mode, cache hit:  0.0887 s (  3.3047 %) ( 30.2603 x)
With ccache, direct mode, cache miss:       2.7757 s (103.4676 %) (  0.9665 x)
With ccache, direct mode, cache hit:        0.0123 s (  0.4595 %) (217.6293 x)
With ccache, depend mode, cache miss:       2.7122 s (101.1020 %) (  0.9891 x)
With ccache, depend mode, cache hit:        0.0124 s (  0.4609 %) (216.9495 x)
```