@erijo 2020-02-25
I've updated BLAKE3 to the latest release BLAKE3-team/BLAKE3@c-0.2.2 and rebased on latest ccache master.

Updated performance numbers now look like this:
```
Without ccache:                             3.1337 s (100.0000 %) (  1.0000 x)
With ccache, preprocessor mode, cache miss: 3.2211 s (102.7911 %) (  0.9728 x)
With ccache, preprocessor mode, cache hit:  0.0899 s (  2.8695 %) ( 34.8487 x)
With ccache, direct mode, cache miss:       3.3066 s (105.5195 %) (  0.9477 x)
With ccache, direct mode, cache hit:        0.0130 s (  0.4162 %) (240.2744 x)
With ccache, depend mode, cache miss:       3.3053 s (105.4788 %) (  0.9481 x)
With ccache, depend mode, cache hit:        0.0130 s (  0.4141 %) (241.4907 x)
```

Compared to master (e0f5063b):
```
Without ccache:                             3.0477 s (100.0000 %) (  1.0000 x)
With ccache, preprocessor mode, cache miss: 3.1426 s (103.1121 %) (  0.9698 x)
With ccache, preprocessor mode, cache hit:  0.0898 s (  2.9453 %) ( 33.9520 x)
With ccache, direct mode, cache miss:       3.1478 s (103.2845 %) (  0.9682 x)
With ccache, direct mode, cache hit:        0.0165 s (  0.5406 %) (184.9961 x)
With ccache, depend mode, cache miss:       3.1039 s (101.8438 %) (  0.9819 x)
With ccache, depend mode, cache hit:        0.0165 s (  0.5405 %) (185.0014 x)
```