@erijo 2020-06-17
I've rebased on top of latest master, ported the build to cmake and updated to latest BLAKE3.

Running `../misc/performance /usr/bin/g++ -std=c++11 -I../src -I. ../src/ccache.cpp` in my build directory results in the following when using master (53f8efff):
```
Without ccache:                             1.2119 s (100.0000 %) (  1.0000 x)
With ccache, preprocessor mode, cache miss: 1.3758 s (113.5180 %) (  0.8809 x)
With ccache, preprocessor mode, cache hit:  0.1021 s (  8.4253 %) ( 11.8691 x)
With ccache, direct mode, cache miss:       1.4012 s (115.6167 %) (  0.8649 x)
With ccache, direct mode, cache hit:        0.0159 s (  1.3130 %) ( 76.1634 x)
With ccache, depend mode, cache miss:       1.4119 s (116.4988 %) (  0.8584 x)
With ccache, depend mode, cache hit:        0.0160 s (  1.3173 %) ( 75.9121 x)
```

Compated with this when using this PR:
```
Without ccache:                             1.2554 s (100.0000 %) (  1.0000 x)
With ccache, preprocessor mode, cache miss: 1.4155 s (112.7486 %) (  0.8869 x)
With ccache, preprocessor mode, cache hit:  0.0916 s (  7.2974 %) ( 13.7035 x)
With ccache, direct mode, cache miss:       1.3175 s (104.9410 %) (  0.9529 x)
With ccache, direct mode, cache hit:        0.0124 s (  0.9869 %) (101.3293 x)
With ccache, depend mode, cache miss:       1.3245 s (105.4975 %) (  0.9479 x)
With ccache, depend mode, cache hit:        0.0125 s (  0.9960 %) (100.3981 x)
```